Enable changing opposing accounts via AJAX
This functionality uses a fully-transparent <select> element overlaid on
top of an arrow-style indicator element. This approach is more usable on
mobile devices since the list of transactions shows in a full-screen
dialog just like native select boxes.  I got the idea for this approach
from Amazon's redesigned homepage.

Originally I wanted to base all inline editing controls on <a> elements,
for keyboard usability.  This didn't work with the <select> trick above,
so this commit also moves back to using <span> elements and disables the
CSS link outline again.