Try importing images in import_json_file.py
This *does not work* - the error is "gnucash_data.models.DoesNotExist:
Transaction matching query does not exist." which must mean that the
data for imported transactions is not being committed to the database in
a timely fashion.  Closing and reopening the GnuCash API session does
not help, even with a 5 second delay.

The current workaround is to run the import_json_file.py script then the
import_images_from_json.py script immediately after, which kind of sucks
but at least the second script runs quickly.